+++ src/game/server/baseentity.cpp
@@ -9060,7 +9060,7 @@ void CBaseEntity::InputRemoveOutput( inputdata_t& inputdata )
 				// If our names match, remove
 				if (Matcher_NamesMatch(szOutput, dataDesc->externalName))
 				{
-					CBaseEntityOutput *pOutput = (CBaseEntityOutput *)((int)this + (int)dataDesc->fieldOffset[0]);
+					CBaseEntityOutput *pOutput = (CBaseEntityOutput *)((intp)this + (int)dataDesc->fieldOffset[0]);
 					pOutput->DeleteAllElements();
 				}
 			}
@@ -9122,7 +9122,7 @@ void CBaseEntity::InputReplaceOutput( inputdata_t& inputdata )
 				// If our names match, replace
 				if (Matcher_NamesMatch(szOutput, dataDesc->externalName))
 				{
-					CBaseEntityOutput *pOutput = (CBaseEntityOutput *)((int)this + (int)dataDesc->fieldOffset[0]);
+					CBaseEntityOutput *pOutput = (CBaseEntityOutput *)((intp)this + (int)dataDesc->fieldOffset[0]);
 					const char *szTarget;
 					const char *szInputName;
 					const char *szParam;
